finalizando

+++ src/main/java/br/comexport/avaliacao/entities/ReputationUserEntity.java
@@ -8,7 +8,6 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import javax.persistence.*;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -28,7 +27,7 @@ public class ReputationUserEntity implements Serializable {
     private UserEntity user;
 
     @Column
-    private BigDecimal score;
+    private Double score;
 
     @Column(nullable = false, updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
