+++ src/main/java/br/comexport/avaliacao/entities/VoteAnswerEntity.java
@@ -8,7 +8,6 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 import javax.persistence.*;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -36,7 +35,7 @@ public class VoteAnswerEntity implements Serializable {
     private AnswerEntity answer;
 
     @Column
-    private BigDecimal score;
+    private Double score;
 
     @Column(nullable = false, updatable = false)
     @Temporal(TemporalType.TIMESTAMP)
